op_alloy_rpc_types

Struct OpTransactionRequest

pub struct OpTransactionRequest(/* private fields */);

Builder for OpTypedTransaction.

Implementations

impl OpTransactionRequest

pub const fn from(self, from: Address) -> Self

Sets the from field in the call to the provided address

pub const fn transaction_type(self, transaction_type: u8) -> Self

Sets the transactions type for the transactions.

pub const fn gas_limit(self, gas_limit: u64) -> Self

Sets the gas limit for the transaction.

pub const fn nonce(self, nonce: u64) -> Self

Sets the nonce for the transaction.

pub const fn max_fee_per_gas(self, max_fee_per_gas: u128) -> Self

Sets the maximum fee per gas for the transaction.

pub const fn max_priority_fee_per_gas( self, max_priority_fee_per_gas: u128, ) -> Self

Sets the maximum priority fee per gas for the transaction.

pub const fn to(self, to: Address) -> Self

Sets the recipient address for the transaction.

pub const fn value(self, value: U256) -> Self

Sets the value (amount) for the transaction.

pub fn access_list(self, access_list: AccessList) -> Self

Sets the access list for the transaction.

pub fn input(self, input: TransactionInput) -> Self

Sets the input data for the transaction.

pub fn build_typed_tx(self) -> Result<OpTypedTransaction, Self>

Builds OpTypedTransaction from this builder. See TransactionRequest::build_typed_tx for more info.

Note that EIP-4844 transactions are not supported by Optimism and will be converted into EIP-1559 transactions.